# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Dasmesh Stock Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /production:
    get:
      operationId: listProduction
      tags: [production]
      summary: List production entries
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
        - name: stockItemId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Production entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductionEntry"
    post:
      operationId: createProduction
      tags: [production]
      summary: Create a production entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionEntry"
  /production/{id}:
    get:
      operationId: getProduction
      tags: [production]
      summary: Get a production entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Production entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionEntry"
    patch:
      operationId: updateProduction
      tags: [production]
      summary: Update a production entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionEntry"
    delete:
      operationId: deleteProduction
      tags: [production]
      summary: Delete a production entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # DISPATCH ENTRIES
...
components:
  schemas:
...
    StockItem:
      type: object
      required: [id, itemCode, category, size, length, unit, status]
      properties:
        id:
          type: integer
        itemCode:
          type: string
        category:
          type: string
        size:
          type: string
        length:
          type: string
        unit:
          type: string
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
...
    ProductionEntry:
      type: object
      required: [id, date, stockItemId, quantity]
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        stockItemId:
          type: integer
        stockItem:
          $ref: "#/components/schemas/StockItem"
        quantity:
          type: number
        remarks:
          type: ["string", "null"]
        createdById:
          type: integer
        createdByName:
          type: ["string", "null"]
        updatedById:
          type: ["integer", "null"]
        updatedByName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    ProductionInput:
      type: object
      required: [date, stockItemId, quantity]
      properties:
        date:
          type: string
          format: date
        stockItemId:
          type: integer
        quantity:
          type: number
        remarks:
          type: string
    ProductionUpdate:
      type: object
      properties:
        quantity:
          type: number
        remarks:
          type: string

    # DISPATCH
...